feat: Added checkCards function + a fucntion to flip cards back over if they don't match
Added function to check the cards. If the number of open cards equals 2 and the two cards don't match, wait and call the flip back funtion to iterate over the cards and remove the open and show classes.
Set open cards back to zero then increment the number of moves.